requirements.txt
asgiref==3.2.3
Django==3.0
django-autoslug==1.9.6
django-crispy-forms==1.8.1
django-filter==2.2.0
django-rest-framework==0.1.0
djangorestframework==3.11.0
psycopg2==2.8.4
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
whitenoise==4.1.4
